requirements.txt
# Core ML and NLP libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
tokenizers>=0.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# SHAP and explainability
shap>=0.42.0
lime>=0.2.0.1
interpret>=0.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Data processing
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1
langdetect>=1.0.9

# Fairness and bias detection
aif360>=0.5.0
fairlearn>=0.8.0
fairness-metrics>=0.7.0

# Web framework for dashboard
streamlit>=1.25.0
gradio>=3.35.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
nbformat>=5.9.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
rich>=13.4.0
colorama>=0.4.6

# Data validation
pydantic>=2.0.0
marshmallow>=3.20.0

# Performance monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# Optional: GPU support
# cupy-cuda11x>=12.0.0  # Uncomment for CUDA 11.x support
# cupy-cuda12x>=12.0.0  # Uncomment for CUDA 12.x support 